% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{aplore3}
\alias{aplore3}
\alias{aplore3-package}
\title{Datasets from Hosmer, Lemeshow, Sturdivant "Applied Logistic Regression" (3rd ed.)}
\description{
This package is a 'unofficial' companion to the textbook Applied
Logistic Regression (3rd ed.,2013) by DW. Hosmer, S. Lemeshow and
RX. Sturdivant. It includes all the data sets used in the book, to
make even more quick/simple to reproduce analysis from R.
}

