% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{burn}
\alias{burn}
\title{BURN data}
\format{A data frame with 100 rows and 3 variables:
\describe{
\item{id}{Identification code}
\item{facility}{Burn facility}
\item{death}{Hospital discharge status}
\item{age}{Age at admission (years)}
\item{gender}{Gender}
\item{race}{Race}
\item{tbsa}{Total burn surface area}
\item{inh_inj}{Burn involved inhalation injury}
\item{flame}{Flame involved in burn injury}
}}
\usage{
burn
}
\description{
burn dataset.
}
\examples{
head(burn, n = 10)
}
\keyword{datasets}

