% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{burn1000}
\alias{burn1000}
\title{BURN1000 data}
\format{A data frame with 100 rows and 3 variables:
\describe{
\item{id}{Identification code}
\item{facility}{Burn facility}
\item{death}{Hospital discharge status}
\item{age}{Age at admission (years)}
\item{gender}{Gender}
\item{race}{Race}
\item{tbsa}{Total burn surface area}
\item{inh_inj}{Burn involved inhalation injury}
\item{flame}{Flame involved in burn injury}
}}
\usage{
burn1000
}
\description{
burn1000 dataset.
}
\examples{
head(burn1000, n = 10)
}
\keyword{datasets}

