% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/burn.R
\docType{data}
\name{burn_eval_1}
\alias{burn_eval_1}
\title{BURN_EVAL_1 data}
\format{A data.frame with 500 rows and 9 variables: the covariate are
the same as those from \code{\link{burn1000}}.}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
burn_eval_1
}
\description{
burn_eval_1 dataset.
}
\examples{
head(burn_eval_1, n = 10)
summary(burn_eval_1)
}
\keyword{datasets}

