% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{chdage}
\alias{chdage}
\title{CHDAGE data}
\format{A data frame with 100 rows and 3 variables:
\describe{
   \item{id}{Identification code}
   \item{age}{Age (years)}
   \item{agegrp}{Age group}
   \item{chd}{Presence of CHD}
}}
\usage{
chdage
}
\description{
chdage dataset.
}
\examples{
## Table 1.1
chdage
## Table 1.3
summary( glm( chd ~ age, family = binomial, data = chdage ))
}
\keyword{datasets}

