% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{glow11m}
\alias{glow11m}
\title{GLOW11M data}
\format{A data.frame with 238 rows and 16 variables: the covariate are
the same as those from \code{\link{glow500}} with the addition of
\describe{
\item{pair}{Pair Identification Code (1-119)}
}}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
glow11m
}
\description{
glow11m dataset.
}
\examples{
head(glow11m, n = 10)
summary(glow11m)
}
\keyword{datasets}

