% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glow.R
\docType{data}
\name{glow_bonemed}
\alias{glow_bonemed}
\title{GLOW_BONEMED data}
\format{
A data.frame with 500 rows and 18 variables: the covariate are
the same as those from \code{\link{glow500}} with the addition of
\describe{
\item{bonemed}{Bone medications at enrollment (1: No, 2: Yes)}
\item{bonemed_fu}{Bone medications at follow-up (1: No, 2: Yes)}
\item{bonetreat}{Bone medications both at enrollment and follow-up (1: No,
2: Yes)}
}
}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
glow_bonemed
}
\description{
glow_bonemed dataset.
}
\examples{
head(glow_bonemed, n = 10)
summary(glow_bonemed)
}
\keyword{datasets}
