% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glow.R
\docType{data}
\name{glow_mis_wmissing}
\alias{glow_mis_wmissing}
\title{GLOW_MIS_WMISSING data}
\format{A data.frame with 500 rows and 10 variables: the covariate are
the same as those from \code{\link{glow500}}, without \code{bmi},
\code{premeno}, \code{armassist}, \code{smoke} and \code{fracscore}.}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
glow_mis_wmissing
}
\description{
glow_mis_wmissing dataset.
}
\examples{
head(glow_mis_wmissing, n = 10)
summary(glow_mis_wmissing)
}
\keyword{datasets}

