% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{glow_rand}
\alias{glow_rand}
\title{GLOW_RAND data}
\format{A data.frame with 500 rows and 15 variables: the covariate are
the same as those from \code{\link{glow500}}.}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
glow_rand
}
\description{
glow_rand dataset.
}
\examples{
head(glow_rand, n = 10)
summary(glow_rand)
}
\keyword{datasets}

