% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\docType{data}
\name{scale_example}
\alias{scale_example}
\title{SCALE_EXAMPLE data}
\format{A data.frame with 500 rows and 2 variables:
\describe{
\item{y}{a dicotomic variable (say 1: No, 2: Yes)}
\item{x}{a numeric variable}
}}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
scale_example
}
\description{
scale_example dataset.
}
\examples{
head(scale_example, n = 10)
summary(scale_example)
}
\keyword{datasets}

