% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{burn13m}
\alias{burn13m}
\title{BURN13M data}
\format{A data.frame with 388 rows and 11 variables: the covariate are
the same as those from \code{\link{burn1000}} with the addition of
\describe{
\item{pair}{Pair Identification Code (1-119)}
}}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
burn13m
}
\description{
burn13m dataset.
}
\examples{
head(burn13m, n = 10)
summary(burn13m)
}
\keyword{datasets}

