% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{chdage}
\alias{chdage}
\title{CHDAGE data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
   \item{id}{Identification code}
   \item{age}{Years}
   \item{chd}{Presence of CHD}
}}
\usage{
chdage
}
\description{
chdage dataset.
}
\examples{
summary( glm( chd ~ age, family = binomial, data = chdage ))
}
\keyword{datasets}

