% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glow.R
\docType{data}
\name{glow11m}
\alias{glow11m}
\title{GLOW11M data}
\format{
A data.frame with 238 rows and 16 variables: the covariate are
the same as those from \code{\link{glow500}} with the addition of
\describe{
\item{pair}{Pair Identification Code (1-119)}
}
}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013) Applied
Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
glow11m
}
\description{
glow11m dataset.
}
\examples{
head(glow11m, n = 10)
summary(glow11m)

## Table 7.2 p. 252
library(survival)
mod7.2 <- clogit(as.numeric(fracture) ~ height + weight + bmi +
                 priorfrac + premeno + momfrac + armassist + raterisk +
                 strata(pair), data = glow11m)
summary(mod7.2)
}
\keyword{datasets}
